continued working on flight schedule
Optimized the flight list by not keeping an array on hand.
Solved the other AIrport problem and the generate a flight from DB problem by creating new object to hold plane with quantity/

for tomorow i need to fix generating a plane and binding it to the airline.
Generating the plane from DB should be efficiant and according to the range between two airports, that should be calculated with the haversine formulas